requirements.txt
elasticsearch==8.17.2
loguru==0.7.3
pandas==2.2.3
python-dotenv==1.0.1
streamlit==1.42.2